Use security rules to control Language Team access.
Use the native seam mechanisms to check for permissions to create, update and delete language team members. Add a security rule in security.drl to configure this access.
Add an initial integration test to verify that hibernate level access is being controlled by the security rules. This test is not yet added to the Seam Test suite.
Add security rules that allow language team members to add/modify translations.