Change handling of single empty references and flags in PO files.
Before this revision, no flags/references and a single empty flag/reference would be stored in the same manner in the database (an empty string), and interpreted as the later when pulling or generating a PO file. In this revision, empty tag/references list are handled with a null entry in the database.