rhbz1023258 - Copy trans refactor and bug-fix.
Correct the bug that prevented copy trans from getting the latest translation when multiple translations were reuse candidates. This was due to the fact that copy trans was not using the right column for ordering the results.
Since it was not trivial or efficient to query all translations for a single document, copy trans is now split to do a very quick an efficient query per text flow. This means copy trans for a document will probably take a longer time, but it will also be more scalable.
Also remove some tests that don't make any sense after the changes.
Correct copy trans time remaining and elapsed messages.
Fix JSF issue causing the stats table to show as reloading with every progress bar update.
Add a test that checks that the latest translation is being used when all things are equal.