Fix kerberos from-based authentication.
Using a servlet to generate the login form was causing the NegotiationAuthenticator valve to return the wrong response code, halting the SPNego process and preventing ticket based authentication (if available) from taking effect.

Replaced the servlet with a jsp file which doesn't have the same problem. Static files work too (.html) but there is a need to dynamically generate the redirection url.