Usability fixes.
Includes changes to the autocomplete controller so that it shows ajax indicators as well as resetting to its initial state when it needs to.
When Escape key is hit on the autocomplete, the result list is hidden and the whole text is selected.
When searching for projects, the presence of spaces no longer omits results.
Remove some default Seam messages that conflict with application defined messages.
Version Group page changes:
Refactor the page so that when adding or removing versions, languages and maintainers, there are ajax indicators for these operations.
Make sure the loaded tab (according to the URL) is displayed before sending off the initial ajax requests.