Prevent opening of file select dialog when clicking outside the intended click target.
The dialog was opening when clicking anywhere in a region to the right of and below
the "browse files" text that is the intended click target. This change limits the
target to just the "browse files" text.
Wrapping of the "browse files" text is prevented since the transparent file upload
button must overlay all of the text for the dialog to open properly (plus it looks
better with the whole link on the next line than split across lines).