Add confirmation to all operations that cancel an upload.
Makes sure the event 'hide.zanata.modal' is triggered in all cases that
close the modal, then add the check to that event.
Also incidentally fixed some cases in which the modal could be closed
without refreshing the document list by attaching an event on
'hide.zanata.modal' to refresh the statistics.