Ensure uploading file items display with consistent height.
During multi-file upload, file items were changing size slightly
when they finished uploading, due to the loader widget changing the
height. This is a workaround until the loader widget is updated to
work better in inline mode - it just adds a hidden loader to the
uploaded template to force it to the same height.