Use format strings for UI text
All strings for display are wrapped in i18n function.
At the moment this only replaces placeholders since I have not provided
a way to make translated strings available. The simplest approach I can
think of to make translated strings available would be to output them as
text in hidden DOM elements of the component, since that would allow
translations to be handled through the existing page translation workflow.