rhbz1154555 - Fix process manager screen issue.
The problem arises when there is a process running and the process manager screen is loaded. With the old async task implementation, the process manager expected a task name. This field is no longer present in the async task handles. For the time being, the process will present a string representation of the task handle. This should be changed in future revisions.