Replace page getDriver.findElement behaviour with search and wait
In yearning for stability, we want to always endure the elements are
there before we attempt to interact with them. There are two levels of
this:
- The element exists and its attributes can be read
- The element is displayed and enabled, ready for interaction.

As much as possible, this change replaces the unchecked interactions
with search and wait, and removes the existing inline waitFor.untils
that were introduced in specific problem areas.