Ensure list operations are visible when appropriate.
It was possible to have a language selected in one of the language
lists, but not see the available operations. This was caused when
the lists and buttons are updated by ajax but the button visibility
code has not run. Fixed by triggering a change event on the two
lists, which causes button visibility to be evaluated and updated.