Add interface for language list settings backend.
The interface is not strictly necessary at this point, but it provides
a compile-time check that an action bean provides everything needed by
the template, rather than having to deal with runtime errors.

The interface can also provide a basis for splitting the language
settings code into separate classes to reduce duplication and move
closer to a single responsibility per class.