Add REST endpoint for version locale list.
Added a common parent for both the project and version locales endpoint
to avoid duplicating the code for formatting the locale list response
entity. This could have been done with a static method on one or the
other class, or on a utility class. I may yet change to one of those
approaches.

Adds a helper method in LocaleServiceImpl to determine whether to get
the project or version locale aliases, to be consistent with how the
locale list is similarly differentiated.