Change resId for plaintext and libreoffice to use content hash.
This includes migration of resId in all existing documents of
the above types to be the content hash.

Conflicts:
	zanata-war/src/main/resources/db/changelogs/db.changelog-3.7.xml

The conflicting file did not yet exist on this branch (this is a back-ported change)
so it was added with only the basic xml structure and the relevant change.

An element was also added in the main db.changelog file to import the conflicting
file.