Keep format strings in the same scope as String.format()
Format strings were being passed to a function, which seems confusing
and so more error-prone. Changed to passing an already-formatted string
so that the format strings can be inlined in the call to String.format()
so that it is obvious that they are format strings and what the
placeholders represent.