Revert "Change token regex for similarity comparison to pick up more cases."
This reverts commit db71115b4b8aba64bcc859f010f4cc8675eff3e2.

More test-cases are required to ensure this regex is correct. Rolling back for
now so that this change does not go into the next release without thorough
testing.

The main problem with the new regex is that it would break on '.' in constructions
such as URLs, which does not appear to be desired behaviour.