Use Groovy script to install appserver
This should make it easier to switch between EAP and WildFly for
testing.  Assuming you have set `EAP6_URL`, and you're happy to use
the version of WildFly specified in the pom, you just switch between
`-Dappserver=jbosseap6` and `-Dappserver=wildfly8`.

No need to specify any of these (and it's easier to switch if you don't):
- cargo.installation
- cargo.basename
- appserver.dir.name

You can still specify these if you want to change the version of WildFly
itself or the version of the management api::
- wildfly.version
- wildfly.client.version