refactor(jsf): match doctype to root element in project.xhtml
A warning was generated by the IDE because the root element and doctype did not
match. We have a mix of doctype "html" and "composition" for <ui:composition>
templates and they seem to work with either, so it seems sensible to make them
consistent by changing them all to the one that does not generate a warning.