refactor(HLocale): add note to change method names to be more standard.
There are a bunch of methods named `retrieve...` rather than `get...`,
which means they don't work with normal property syntax. There are normal
getters, but they may return nothing because the fields are populated only
in the `retireve` methods.
This needs a bit of thought and attention to make it sensible so it does
not waste developer time.