fix(PermissionGranter): allow proxies to match accepted types
The target object for permission was having its class checked using an == comparison,
which returned false for proxy objects. The check is changed to isAssignableFrom so
that any valid subclass (including a proxy) will satisfy the check.
This does mean that if any of the entities we check security for is a subclass of
another entity we check security for, there is potential for a false-positive match,
but none of our entities inherit in that way.