fix(security): do not output every user email address in the DOM
User email address was being output as the id of a DOM element, which is easily seen
by anyone with a browser. It looks like the id was not even being used anywhere.
I do not know why this was done. Was it an attempt to generate a unique ID for each
list item? The same person can be output in multiple categories so it would not have
been unique anyway. I just removed it, and the list still seems to work properly.