fix: scroll permission modal to top whenever it opens
This is also fixed in assets, but there may be some issues with updating assets version
at the moment so I chose to add a safe workaround for now. There will be no harm when
the assets version is updated, but this change can be reverted when assets is stable and
updated.