fix(ZNTA-1067): fix kerberos authentication
This change solves a problem where Kerberos authentication doesn't get
triggered due to the new Url rewriter changing the expected response
code and thus stopping the whole process. The old klogin.xhtml page is
replaced with a servlet which performs the same tasks without the need
to go through the Url rewriter.