chore(frontend): update test dependencies and configure tests
Tests were not configured properly and the modules under test had diverged.
It looks like tests have not run for a long time.
This fixes test config so tests can run with `npm test` and all the non-broken
tests will run and pass. Some tests need to be updated due to module changes.